Changed js code to tsx. Added comments. Fixed types
