+++ client/package.json
@@ -18,7 +18,6 @@
     "adventure-component-library": "0.0.24",
     "animejs": "^3.2.0",
     "dnd-multi-backend": "^6.0.0",
-    "emoji-data": "^0.2.0",
     "emoji-picker-react": "^3.2.1",
     "firebase": "^7.16.0",
     "immutability-helper": "^3.0.2",
@@ -59,6 +58,7 @@
     ]
   },
   "devDependencies": {
+    "@types/animejs": "^3.1.1",
     "@types/react-router-dom": "^5.1.5",
     "@types/uuid": "^8.0.0",
     "gh-pages": "^3.1.0",
