+++ client/src/components/css/Widget.css
@@ -78,17 +78,20 @@
 }
 
 .live_emoji {
-  width: 1.8rem;
-  height: 1.8rem;
+  width: 30px;
+  height: 30px;
   position: absolute;
   left: -2rem;
+  overflow: hidden
 }
 
 #emojis {
-  width: 27rem;
+  width: 100vw;
   height: 15rem;
-  margin-top: auto;
+  margin-top: 1.2rem;
   margin-bottom: auto;
+  align-self: center;
   position:relative;
-  overflow:hidden;
+  overflow: hidden
+
 }
